refactor(core): split facade into submodules
